gerekli guncellemeler

+++ WindowsFormsApp1/AllWindowsForms/CariSatisEkrani.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Windows.Forms;
 using WindowsFormsApp1.Context;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement;
 
 namespace WindowsFormsApp1
 {
@@ -98,24 +99,40 @@ namespace WindowsFormsApp1
 
         private void button2_Click(object sender, EventArgs e)
         {
+
             int ID = int.Parse(dataGridView3.CurrentRow.Cells[0].Value.ToString());
             var tbl = dbContext.Satis.FirstOrDefault(x => x.SatisID == ID);
 
             int ID3 = int.Parse(txtKod.Text);
             var tbl2 = dbContext.Uruns.FirstOrDefault(x => x.BarkodNo == ID3);
-            tbl2.StokMiktar += int.Parse(txtSatis.Text);
-
             int ID2 = int.Parse(txtMus.Text);
             var tbl3 = dbContext.Musteris.FirstOrDefault(x => x.MusteriID == ID2);
-            tbl3.MusteriBorc -= ((decimal.Parse(txtSatis.Text)) * (tbl2.SatisFiyati));
-            tbl3.GuncelBorc = tbl3.MusteriBorc - tbl3.MusteriOdeme;
 
-            dbContext.Satis.Remove(tbl);
-            dbContext.SaveChanges();
-            MessageBox.Show("Sepetten silindi");
-            dataGridView1.DataSource = dbContext.Uruns.ToList();
-            dataGridView2.DataSource = dbContext.Musteris.ToList();
-            dataGridView3.DataSource = dbContext.Satis.ToList();
+            string kullaniciAdi = textBox2.Text;
+            var x1 = dbContext.Kullanicis.FirstOrDefault(x => x.Ad == kullaniciAdi);
+            string kullanicisifre = textBox1.Text;
+            var x2 = dbContext.Kullanicis.FirstOrDefault(x => x.Sifre == kullanicisifre);
+            if (x1==null||x2==null)
+            {
+                MessageBox.Show("Lütfen geçerli bir kullanıcı girin");
+
+            }
+            else
+            {
+                tbl2.StokMiktar += int.Parse(txtSatis.Text);
+
+                
+                tbl3.MusteriBorc -= ((decimal.Parse(txtSatis.Text)) * (tbl2.SatisFiyati));
+                tbl3.GuncelBorc = tbl3.MusteriBorc - tbl3.MusteriOdeme;
+
+                dbContext.Satis.Remove(tbl);
+                dbContext.SaveChanges();
+                MessageBox.Show("Sepetten silindi");
+                dataGridView1.DataSource = dbContext.Uruns.ToList();
+                dataGridView2.DataSource = dbContext.Musteris.ToList();
+                dataGridView3.DataSource = dbContext.Satis.ToList();
+            }
+            
         }
     }
 }
